Calculating Time Interval Statistics
We determined the Time Interval Statistics to decide on the optimal or Threshold Time Interval since the data is not recorded consistently. Although the dataset description states that the time interval between two consecutive recordings is 5s, it is not true. There is a little inconsistency in Data Collection and to address this we had to understand the statistics of it.